requirements.txt
pyactr~=0.3.2
numpy~=2.1.3
pandas
matplotlib
seaborn
openai~=1.54.4
scikit-learn~=1.5.2
python-dotenv~=1.0.1
anthropic~=0.39.0